Add files via upload
This is my first project which I worked on building the Slack.com where user can interact  and message in teams. this 6days journey is filled with some headaches and also happiness by overcoming that challenges. But in end I left the project in middle for some personal reason i am not able to do it for mobile version or mode. I put my effort to build this project. I  Learnt some topics like Handling and use of css framework Tailwind Css. By using this my work is reduced and works smoothly.